bramble: move obj/type string lookup to cansnoop
Problem: canmsg_v2.[ch] now only contain functions used
by cansnoop.

Make those functions local to cansnoop and drop that
source module.
