--- bramble/src/libbramble/canmsg_v2.h
@@ -1,16 +1,0 @@
-/* SPDX-License-Identifier: GPL-3.0-or-later */
-
-#ifndef _BRAMBLE_CANMSG_V2_H
-#define _BRAMBLE_CANMSG_V2_H
-
-#include "canmsg.h"
-
-const char *canmsg_objstr (const struct canmsg *msg);
-const char *canmsg_typestr (const struct canmsg *msg);
-
-
-#endif /* !_BRAMBLE_CANMSG_V2_H */
-
-/*
- * vi:ts=4 sw=4 expandtab
- */
